% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCell_transformation.R
\name{AUCell_transformation}
\alias{AUCell_transformation}
\title{Title}
\usage{
AUCell_transformation(sc_matrix, up_signature_list, down_signature_list)
}
\arguments{
\item{sc_matrix}{Single-cell transcriptome expression profile}

\item{up_signature_list}{The list of Gene sets up-regulated by drug instances}

\item{down_signature_list}{The list of Gene sets down-regulated by drug instances}
}
\value{
The drug-by-cell matrix
}
\description{
Title
}
