% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_lgr_drug_score.R
\name{cal_lgr_drug_score}
\alias{cal_lgr_drug_score}
\title{Title}
\usage{
cal_lgr_drug_score(D2C_matrix, cell_labels)
}
\arguments{
\item{D2C_matrix}{The drug-by-cell matrix generated by DrugReSC}

\item{cell_labels}{The cell phenotype labels}
}
\value{
The raw drug scores
}
\description{
Title
}
