% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugReSC.R
\name{DrugReSC}
\alias{DrugReSC}
\title{Title}
\usage{
DrugReSC(
  sc_matrix,
  drug_signature_list,
  PACSI_result,
  D2C_transformation = "ssgsea",
  improtance_method = "rf",
  ncore = 8
)
}
\arguments{
\item{sc_matrix}{Single-cell transcriptome expression profile}

\item{drug_signature_list}{signature list of drug instances}

\item{PACSI_result}{Results of the PACSI method}

\item{D2C_transformation}{Methods for transforming gene-by-cell matrices into drug-by-cell matrices}

\item{improtance_method}{Feature importance calculation method}

\item{ncore}{The number of parallel threads}
}
\value{
Ranked list of candidate drugs
}
\description{
Title
}
