% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drug_iLINCS_sig.R
\name{get_drug_iLINCS_sig}
\alias{get_drug_iLINCS_sig}
\title{Title}
\usage{
get_drug_iLINCS_sig(true_drug_sig_id_list)
}
\arguments{
\item{true_drug_sig_id_list}{ID of the drug instance in the iLINCS database}
}
\value{
Results of drug differences analysis extracted from the iLINCS database
}
\description{
Title
}
