% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drug_sig.R
\name{get_drug_sig}
\alias{get_drug_sig}
\title{Title}
\usage{
get_drug_sig(iLINCS_signature_list, pvalue_cutoff = 0.01, logFC_cutoff = 0)
}
\arguments{
\item{iLINCS_signature_list}{Results of drug differences analysis extracted from the iLINCS database}

\item{pvalue_cutoff}{Thresholds for p values}

\item{logFC_cutoff}{Thresholds for logFC}
}
\value{
signature list of drug instances
}
\description{
Title
}
