% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_normalize.R
\name{min_max_normalize}
\alias{min_max_normalize}
\title{Title}
\usage{
min_max_normalize(x)
}
\arguments{
\item{x}{The raw drug score vector}
}
\value{
The normalized drug score vectors
}
\description{
Title
}
