% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniPath_transformation.R
\name{UniPath_transformation}
\alias{UniPath_transformation}
\title{Title}
\usage{
UniPath_transformation(
  sc_matrix,
  up_signature_list,
  down_signature_list,
  ncore
)
}
\arguments{
\item{sc_matrix}{Single-cell transcriptome expression profile}

\item{up_signature_list}{The list of Gene sets up-regulated by drug instances}

\item{down_signature_list}{The list of Gene sets down-regulated by drug instances}

\item{ncore}{The number of parallel threads}
}
\value{
The drug-by-cell matrix
}
\description{
Title
}
